Re-add baseurl lines, but using example domain
This sorta reverts f30862ce, because it turns out various people
(including me!) actually rely on these commented-out baseurl
lines in various ways. I (in openQA tests) and at least one other
person (hiredman on IRC), for e.g., use them as a base to modify
by changing the host, but keeping the path component, to use a
specific host that is known to always be a fast local one in a
particular environment and avoid the overhead of the metalink/
mirrorlist system.

To avoid encouraging the use of the download.fedoraproject.org
redirect system (as was the original intent), we replace the
hostname with an example one, so the line clearly cannot be used
as-is just by removing the #. We *could* also add a couple of
comment lines specifically discouraging people from using baseurl
unless they're very sure it's what they need, but I don't know
if it's necessary.

Signed-off-by: Adam Williamson <awilliam@redhat.com>