Enable the updates archive repo on non-rawhide.
The fedora-repos-archive RPM is only installed by default in places
where it is needed (i.e. OSTree based distributions). The yum repo
file delivered by the RPM should default to being enabled when on
a numbered version of Fedora (non-rawhide). This was the original
intention when I added it a while back.